+++ pymince/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
 __title__ = "pymince"
-__version__ = "2.11.4"
+__version__ = "2.12.0"